+++ Funcky.Analyzers/Funcky.Analyzers/AlternativeMonadAnalyzer/AlternativeMonadAnalyzer.SelectMany.cs
@@ -1,5 +1,6 @@
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.Operations;
+using static Funcky.Analyzers.AlternativeMonadErrorStateConstructorMatching;
 using static Funcky.Analyzers.FunckyWellKnownMemberNames;
 
 namespace Funcky.Analyzers;
@@ -19,26 +20,4 @@ public partial class AlternativeMonadAnalyzer
     private static bool IsSelectManyEquivalent(AlternativeMonadType alternativeMonadType, IInvocationOperation matchInvocation, INamedTypeSymbol receiverType, IArgumentOperation errorStateArgument)
         => SymbolEqualityComparer.IncludeNullability.Equals(receiverType, matchInvocation.Type)
             && IsErrorStateConstructorReference(alternativeMonadType, errorStateArgument.Value);
-
-    private static bool IsErrorStateConstructorReference(AlternativeMonadType alternativeMonadType, IOperation operation)
-        => operation switch
-        {
-            IPropertyReferenceOperation propertyReference => IsErrorStateConstructorReference(alternativeMonadType, propertyReference),
-            IDelegateCreationOperation { Target: IMethodReferenceOperation { Method: var method } } => IsErrorStateConstructorReference(alternativeMonadType, method),
-            _ => false,
-        };
-
-    private static bool IsErrorStateConstructorReference(AlternativeMonadType alternativeMonadType, IMethodSymbol method)
-        => method is { Name: var name, IsStatic: true, ContainingType: var containingType }
-           && name == alternativeMonadType.ErrorStateConstructorName
-           && IsTypeContainingConstructors(alternativeMonadType, containingType);
-
-    private static bool IsErrorStateConstructorReference(AlternativeMonadType alternativeMonadType, IPropertyReferenceOperation operation)
-        => operation is { Property: { Name: var name, IsStatic: true, ContainingType: var containingType } }
-            && name == alternativeMonadType.ErrorStateConstructorName
-            && IsTypeContainingConstructors(alternativeMonadType, containingType);
-
-    private static bool IsTypeContainingConstructors(AlternativeMonadType alternativeMonadType, INamedTypeSymbol type)
-        => SymbolEqualityComparer.Default.Equals(type.ConstructedFrom, alternativeMonadType.Type)
-            || SymbolEqualityComparer.Default.Equals(type.ConstructedFrom, alternativeMonadType.ConstructorsType);
 }
